gpt.ts back under its file-size budget by moving vision spend policy to its owner
server/gpt.ts was 1476 lines against a 1450 budget. The 53 lines that came out
are the vision block: VisionUseCase, SubscriptionTier, VisionModelDecision,
selectVisionModel and estimateVisionCostUSD.

They were never gpt.ts's job. gpt.ts owns TALKING to the model — building
context, choosing a chat model, calling the API, counting cost. It never called
selectVisionModel or estimateVisionCostUSD once, and neither of them referenced
anything in gpt.ts. The block sat in the file because it mentions model names,
not because it belonged to it.

What the pair actually answers is a cost-to-serve question: what is this member
entitled to spend on a photo (inactive_tier_blocked / food_trial / food_paid),
and what does that decision cost us. server/cost-tracking.ts already owns
exactly that question for the other expensive operation — voiceCostUsd sits
immediately above where these now land, and the file's own header names
"voice + vision" as the two halves of a whale. estimateVisionCostUSD takes a
VisionModelDecision as its input, so splitting the price from the decision that
fixes it would give two files a say in one number.

No new module: the modules counter is at 239/239 with zero headroom, so this
had to go to an existing owner or not happen. It went to the owner that was
already there.

Behaviour is unchanged and provably so — the moved block is byte-identical to
the block on main. The consumers only changed where they import from:
  - handlers/media.ts already imported recordServiceCost from ../cost-tracking,
    so the two names joined that line and the ../gpt import line went away (the
    file drops one line, and it is over budget too).
  - routes.ts imported both names and used neither; the dead names are gone.

Guards after: server/gpt.ts 1423/1450 — off the failing list. The four other
file-size reds and the three architecture reds are unchanged pre-existing
budget breaches, not touched here. tsc clean. No budget raised, nothing
suppressed.
