+++ server/gpt.ts
@@ -604,59 +604,6 @@ export function selectModel(instruction: string, userMessage: string): { model: 
   return { model: "gpt-4o-mini", maxTokens: 160, reason: "coaching" };
 }
 
-export type VisionUseCase = "food_photo" | "progress_compare" | "exercise_classify" | "step_ocr";
-export type SubscriptionTier = "active" | "trial" | "inactive" | string | null | undefined;
-
-export interface VisionModelDecision {
-  allowed: boolean;
-  model: "gpt-4o" | "gpt-4o-mini";
-  detail: "low" | "auto" | "high";
-  maxTokens: number;
-  reason: string;
-}
-
-export function selectVisionModel(useCase: VisionUseCase, tier: SubscriptionTier): VisionModelDecision {
-  const t = (tier || "").toLowerCase();
-  const paying = t === "active";
-  const onboarded = paying || t === "trial";
-
-  if (!onboarded) {
-    return {
-      allowed: false,
-      model: "gpt-4o-mini",
-      detail: "low",
-      maxTokens: 0,
-      reason: "inactive_tier_blocked",
-    };
-  }
-
-  switch (useCase) {
-    case "progress_compare":
-      return paying
-        ? { allowed: true, model: "gpt-4o", detail: "auto", maxTokens: 400, reason: "progress_paid" }
-        : { allowed: true, model: "gpt-4o-mini", detail: "auto", maxTokens: 350, reason: "progress_trial" };
-
-    case "food_photo":
-      // gpt-4o for paying users — better SA food recognition, fewer "cannot make out" failures
-      return paying
-        ? { allowed: true, model: "gpt-4o", detail: "auto", maxTokens: 400, reason: "food_paid" }
-        : { allowed: true, model: "gpt-4o-mini", detail: "auto", maxTokens: 400, reason: "food_trial" };
-
-    case "exercise_classify":
-    case "step_ocr":
-      return { allowed: true, model: "gpt-4o-mini", detail: "low", maxTokens: useCase === "step_ocr" ? 50 : 8, reason: useCase };
-  }
-}
-
-export function estimateVisionCostUSD(decision: VisionModelDecision, completionTokens: number = 0): number {
-  const imgTok = decision.detail === "low" ? 85 : decision.detail === "high" ? 400 : 170;
-  const promptTok = imgTok + 300;
-  if (decision.model === "gpt-4o-mini") {
-    return (promptTok * 0.15 + completionTokens * 0.6) / 1_000_000;
-  }
-  return (promptTok * 5 + completionTokens * 15) / 1_000_000;
-}
-
 export interface GptFoodItem {
   name: string;
   kcal: number;
