+++ server/handlers/media.ts
@@ -28,7 +28,6 @@ import { handleWeightLog } from "./weight";
 import { handleWater } from "./water";
 import { recomputeTodayFoodTotals, scanForSAFoods, findDuplicateMealToday, parseFoodLogTotalsFromMessageOut } from "./food-scanner";
 import { buildFoodVisionSystemPrompt, buildFoodVisionUserPrompt, buildMenuPickPrompt } from "./food-vision-prompt";
-import { selectVisionModel, estimateVisionCostUSD } from "../gpt";
 import { calculateTargets, getDailyStepContext, waterTargetLitres } from "../targets";
 import { buildPhysiqueAnalysisPrompt, parsePhysiqueAnalysis, formatPhysiqueFocusLine, buildProgressComparisonPrompt } from "../physique-analysis";
 import { buildFormCheckPrompt, extractFormExercise } from "../form-check-prompt";
@@ -55,7 +54,7 @@ import { commitFoodLog } from "./food-context";
 import { itemsFromVisionText } from "../serving-units";
 import { tryLogDistanceScreenshot } from "./distance-log";
 import { encodePendingFood } from "../food-referent";
-import { recordServiceCost } from "../cost-tracking";
+import { recordServiceCost, selectVisionModel, estimateVisionCostUSD } from "../cost-tracking";
 import { allowExpensiveOp } from "../usage-governor";
 import { scribeTranscribe } from "../elevenlabs";
 import { extractVideoFrames } from "../video-frames";
